refactor(app): remove dead code form old templates
