+++ app/package.json
@@ -25,7 +25,6 @@
     "i18next": "^22.5.1",
     "i18next-browser-languagedetector": "^7.2.0",
     "i18next-http-backend": "^2.4.2",
-    "js-cookie": "^3.0.5",
     "json5": "^2.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
