--- app/src/components/assetInventory/CardView/index.tsx
@@ -1,1 +1,0 @@
-export * from "./CardView";
