--- app/src/components/assetInventory/ListView/index.tsx
@@ -1,1 +1,0 @@
-export * from "./ListView";
