--- app/src/components/assetInventory/index.ts
@@ -1,2 +1,0 @@
-export * from "./CardView";
-export * from "./ListView";
