--- app/src/components/detailsView/KPIs/index.tsx
@@ -1,1 +1,0 @@
-export * from "./KPIs";
