--- app/src/components/detailsView/Kpi1/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Kpi1";
