--- app/src/components/detailsView/Kpi1/styles.ts
@@ -1,11 +1,0 @@
-import { css } from "@emotion/css";
-import { theme } from "@hitachivantara/uikit-styles";
-
-const styles = {
-  indicator: css({
-    marginLeft: theme.space.sm,
-    marginRight: theme.space.sm,
-  }),
-};
-
-export default styles;
