--- app/src/components/detailsView/Kpi2/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Kpi2";
