--- app/src/components/detailsView/Kpi3/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Kpi3";
