--- app/src/components/detailsView/Table/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Table";
