--- app/src/components/detailsView/Table/styles.ts
@@ -1,10 +1,0 @@
-import { css } from "@emotion/css";
-import { theme } from "@hitachivantara/uikit-styles";
-
-const styles = {
-  section: css({
-    marginBottom: theme.space.md,
-  }),
-};
-
-export default styles;
