--- app/src/components/detailsView/index.ts
@@ -1,6 +1,0 @@
-export * from "./Kpi1";
-export * from "./Kpi2";
-export * from "./Kpi3";
-export * from "./KPIs";
-export * from "./Properties";
-export * from "./Table";
