--- app/src/components/listView/Table/index.ts
@@ -1,1 +1,0 @@
-export * from "./Table";
